refactor(encoding): isolate dict-prime lifecycle in a child module
Convert match_generator.rs into a directory module and move the seven
dictionary entry points (resident check + reapply, prime, snapshot
restore/capture, invalidate, entropy seed) out of the driver's main
`Matcher` impl into a `dict_prime` child module.

The snapshot path reads/writes the driver's most invariant-critical
private state (`primed`, `reset_shape`, the `PrimedKey` snapshot identity
that guards against restoring a mismatched matcher shape and emitting an
undecodable match). A child module reaches those private fields directly,
so the concern relocates with zero field widening -- encapsulation stays
intact. The trait surface keeps thin `#[inline]` forwarders to the child's
inherent `*_impl` helpers, so output is byte-identical.

match_generator/mod.rs: 7438 -> 7000.

Part of #441.
